# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations": {
      "get": {
        "operationId": "OrganizationController_listTenants",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TenantSummaryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/administrative-divisions": {
      "get": {
        "operationId": "OrganizationController_administrativeDivisions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdministrativeDivisionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/enterprise-nature-tags": {
      "get": {
        "operationId": "OrganizationController_enterpriseNatureTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnterpriseNatureTagDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/industry-categories": {
      "get": {
        "operationId": "OrganizationController_industryCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IndustryCategoryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/current": {
      "get": {
        "operationId": "OrganizationController_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentOrganizationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/universities": {
      "post": {
        "operationId": "OrganizationController_createUniversity",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUniversityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedTenantDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/universities/{tenantId}/admins": {
      "post": {
        "operationId": "OrganizationController_addUniversityAdmin",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitialTenantAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedUniversityAdminDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/enterprises": {
      "post": {
        "operationId": "OrganizationController_createEnterprise",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnterpriseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedTenantDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/enterprises/{tenantId}/admins": {
      "post": {
        "operationId": "OrganizationController_addEnterpriseAdmin",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitialTenantAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedUniversityAdminDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/governments": {
      "post": {
        "operationId": "OrganizationController_createGovernment",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGovernmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedTenantDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/governments/{tenantId}/admins": {
      "post": {
        "operationId": "OrganizationController_addGovernmentAdmin",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitialTenantAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedUniversityAdminDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
    "/api/v1/organizations/governments/{tenantId}/university-scopes": {
      "put": {
        "operationId": "OrganizationController_replaceGovernmentScopes",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceGovernmentScopesDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Organization"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zhixing Hub API",
    "description": "知行工坊模块化单体业务接口",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zhixing_session"
      }
    },
    "schemas": {
...
      "TenantSummaryDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "UNIVERSITY",
              "ENTERPRISE",
              "GOVERNMENT"
            ]
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          },
          "isPublicAcademy": {
            "type": "boolean",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "tenantId",
          "type",
          "name",
          "status",
          "isPublicAcademy",
          "createdAt"
        ]
      },
      "AdministrativeDivisionDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": "string",
            "enum": [
              "PROVINCE",
              "PREFECTURE"
            ]
          },
          "parentCode": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "code",
          "name",
          "level",
          "parentCode"
        ]
      },
      "EnterpriseNatureTagDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "code",
          "name",
          "sortOrder"
        ]
      },
      "IndustryCategoryDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "CurrentOrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "PLATFORM",
              "UNIVERSITY",
              "ENTERPRISE",
              "GOVERNMENT"
            ]
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "status"
        ]
      },
      "InitialTenantAdminDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "description": "用户端管理员手机号；租户管理员不是平台管理账号",
            "example": "13800138000"
          }
        },
        "required": [
          "displayName",
          "phone"
        ]
      },
      "CreateUniversityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "initialAdmin": {
            "$ref": "#/components/schemas/InitialTenantAdminDto"
          }
        },
        "required": [
          "name",
          "initialAdmin"
        ]
      },
      "CreatedTenantDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "UNIVERSITY",
              "ENTERPRISE",
              "GOVERNMENT"
            ]
          },
          "name": {
            "type": "string"
          },
          "initialAdminAccountId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "tenantId",
          "type",
          "name",
          "initialAdminAccountId"
        ]
      },
      "CreatedUniversityAdminDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "accountId"
        ]
      },
      "CreateEnterpriseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "initialAdmin": {
            "$ref": "#/components/schemas/InitialTenantAdminDto"
          },
          "natureTagCode": {
            "type": "string",
            "example": "GENERAL"
          },
          "industryCategoryCode": {
            "type": "string",
            "example": "I"
          },
          "locationCodes": {
            "description": "至少一个受控地级市行政区划代码",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "initialAdmin",
          "natureTagCode",
          "industryCategoryCode",
          "locationCodes"
        ]
      },
      "CreateGovernmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "initialAdmin": {
            "$ref": "#/components/schemas/InitialTenantAdminDto"
          },
          "divisionCode": {
            "type": "string"
          },
          "visibleUniversityTenantIds": {
            "description": "政务端允许聚合查看的高校租户白名单",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "initialAdmin",
          "divisionCode",
          "visibleUniversityTenantIds"
        ]
      },
      "ReplaceGovernmentScopesDto": {
        "type": "object",
        "properties": {
          "visibleUniversityTenantIds": {
            "description": "政务端允许聚合查看的高校租户白名单",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "visibleUniversityTenantIds"
        ]
      },
...
    }
  }
}